requirements.txt
# Core dependencies
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.7
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
requests==2.32.4

# ML and NLP
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.7.0
spacy==3.8.7
transformers==4.52.4
torch==2.2.2
torchvision==0.17.2
sentence-transformers==4.1.0
nltk==3.8.1
gensim==4.3.3
openai-whisper==20231117

# Text processing
beautifulsoup4==4.12.2
lxml==5.4.0
html5lib==1.1
markdown==3.8
python-docx==1.0.1
PyPDF2==3.0.1
pytesseract==0.3.10

# Image processing
Pillow==10.4.0
opencv-python==4.11.0.86
imageio==2.37.0

# Audio/Video processing
moviepy==2.2.1
pydub==0.25.1
librosa==0.11.0

# Database
sqlalchemy==2.0.41
alembic==1.16.2
psycopg2-binary==2.9.10

# Testing
pytest==7.4.3
pytest-asyncio==0.23.8
pytest-cov==4.1.0
pytest-mock==3.14.1
httpx==0.28.1

# Development
black==25.1.0
flake8==7.2.0
mypy==1.16.0
isort==6.0.1
pre-commit==4.2.0

# Documentation
mkdocs==1.6.1
mkdocs-material==9.6.14
mkdocstrings==0.29.1

# Monitoring and logging
prometheus-client==0.22.1
python-json-logger==3.3.0
sentry-sdk==2.30.0

# Security
python-dotenv==1.1.0
cryptography==45.0.4

# API clients
google-api-python-client==2.172.0
tweepy==4.15.0
praw==7.8.1
medium-api==0.6.0
rdflib==7.1.4
wikidata==0.8.1
wikipedia==1.4.0

# Additional dependencies
anthropic==0.54.0
gradio==3.50.0
pandas==2.3.0
tqdm==4.67.1
soundfile==0.13.1
numba==0.61.2
python-magic==0.4.27
python-magic-bin==0.4.14
jsonschema==4.24.0 